Address comments and make node pool deletion delay configurable to allow tests to pass
